requirements.txt
streamlit
instructor
openai
beautifulsoup4
requests
python-dotenv
pydantic
tavily-python
validators